Add day2 part 1 solution
Yeah I know this contains duplicate code in the runOpcodes
function and could be refactored later to be improved. I'll
leave it as is for now though 'cause it got me the answer.

This solution is pretty simple and easy to understand. Scan
the file for the input, then split it by commas to an array (or
in case of Go, slice). Then convert each string element into
an int and add to a new int array to be passed to the "opcode"
function that will run the requirements of the puzzle.

I know I could use another function to scan for the input here
but I wanted to reuse the code I used in day1.