Add part2 solution
Unlike with part1, this one doesn't use the runOpcodes
function but rather a simple for loop to "run" the "Intcode
program". And instead of a switch block, I used an if elseif
instead to make things simpler.

The makeIntcodes function is used to generate an int slice
of the input. This is to make sure that the value of the `ic`
resets before every inner loop.